third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-not-below.html

<!DOCTYPE html>
<head>
<script src="../../resources/js-test.js"></script>
<script src="resources/spatial-navigation-utils.js"></script>
</head>
<style>
body a {
    position: absolute;
    height: 100px;
    width: 100px;
    border: 1px solid black;
}
#e1 {
    left: 10px;
    top: 210px;
}
#e2 {
    left: 120px;
    top: 250px;
}
#e3 {
    left: 250px;
    top: 250px;
    height: 120px;
}
</style>
<body id="some-content" onload="runTest()">
<p id="description"></p>
<div id="elements">
<a id="e1" href="#e1">Element 1</a>
<a id="e2" href="#e2">Element 2</a>
<a id="e3" href="#e3">Element 3</a>
</div>
<div id="console"></div>
<script type="application/javascript">
description('This test ensures that e2 is NOT below e1 and e3 is NOT below e2.');

jsTestIsAsync = true;

var resultMap = [
    ["Down", "e1"],
    ["Right", "e2"],
    ["Down", "e2"],
    ["Right", "e3"],
    ["Up", "e3"],
    ["Left", "e2"],
    ["Up", "e2"],
    ["Left", "e1"],
    ["DONE", "DONE"]
];

if (window.testRunner) {
    testRunner.overridePreference("WebKitTabToLinksPreferenceKey", 1);
    testRunner.overridePreference("WebKitSpatialNavigationEnabled", 1);
}

function runTest()
{
    // starting the test itself: get to a known place.
    document.getElementById("e1").focus();

    initTest(resultMap, finishJSTest);
}
</script>
</body>
</html>
