third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-not-below-expected.txt

This test ensures that e2 is NOT below e1 and e3 is NOT below e2.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Element 1Element 2Element 3
PASS gFocusedDocument.activeElement.getAttribute("id") is "e1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e1"
PASS successfullyParsed is true

TEST COMPLETE
